+++ src/test/java/com/ing/data/cassandra/jdbc/codec/SmallintToIntCodecTest.java
@@ -1,0 +1,89 @@
+/*
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
+package com.ing.data.cassandra.jdbc.codec;
+
+import com.datastax.oss.driver.api.core.ProtocolVersion;
+import com.datastax.oss.driver.api.core.type.DataTypes;
+import com.datastax.oss.driver.api.core.type.reflect.GenericType;
+import org.apache.commons.lang3.StringUtils;
+import org.junit.jupiter.api.Test;
+
+import java.nio.ByteBuffer;
+
+import static com.ing.data.cassandra.jdbc.Utils.NULL_KEYWORD;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNull;
+
+public class SmallintToIntCodecTest {
+
+    private final SmallintToIntCodec sut = new SmallintToIntCodec();
+
+    @Test
+    void givenCodec_whenGetJavaType_returnInt() {
+        assertEquals(GenericType.INTEGER, sut.getJavaType());
+    }
+
+    @Test
+    void givenCodec_whenGetCqlType_returnVarint() {
+        assertEquals(DataTypes.SMALLINT, sut.getCqlType());
+    }
+
+    @Test
+    void givenNullValue_whenEncode_returnNull() {
+        assertNull(sut.encode(null, ProtocolVersion.DEFAULT));
+    }
+
+    @Test
+    void givenValue_whenEncode_returnByteBuffer() {
+        ByteBuffer bytes = sut.encode(12345, ProtocolVersion.DEFAULT);
+        assertNotNull(bytes);
+        assertEquals(12345, bytes.getInt());
+    }
+
+    @Test
+    void givenNullValue_whenDecode_returnNull() {
+        assertNull(sut.decode(null, ProtocolVersion.DEFAULT));
+    }
+
+    @Test
+    void givenValue_whenDecode_returnInt() {
+        ByteBuffer bytes = ByteBuffer.allocate(2).putShort((short) 12345);
+        bytes.position(0);
+        assertEquals(12345, sut.decode(bytes, ProtocolVersion.DEFAULT));
+    }
+
+    @Test
+    void givenNullOrEmptyValue_whenParse_returnNull() {
+        assertNull(sut.parse(null));
+        assertNull(sut.parse(NULL_KEYWORD));
+        assertNull(sut.parse(StringUtils.EMPTY));
+        assertNull(sut.parse(StringUtils.SPACE));
+    }
+
+    @Test
+    void givenNonNullValue_whenParse_returnExpectedValue() {
+        assertEquals(12345, sut.parse("12345"));
+    }
+
+    @Test
+    void givenNullValue_whenFormat_returnNull() {
+        assertEquals(NULL_KEYWORD, sut.format(null));
+    }
+
+    @Test
+    void givenNonNullValue_whenFormat_returnExpectedValue() {
+        assertEquals("12345", sut.format(12345));
+    }
+}